Modificado en ventas

+++ resources/views/livewire/ventas.blade.php
@@ -85,14 +85,14 @@
                                                                     </button>
                                                                 @endif
                                                             @else
-                                                                @if($venta->user_id === auth()->id() || auth()->user()->canManageCurrentTenant())
+                                                                @if((int)$venta->user_id === (int)auth()->id() || auth()->user()->canManageCurrentTenant())
                                                                     <a href="{{ route('venta', ['ventaId' => $venta->id]) }}"
                                                                         class="btn btn-sm btn-success"
                                                                         title="Continuar venta">
                                                                         <i class="fa-solid fa-arrow-right"></i>
                                                                     </a>
                                                                 @endif
-                                                                @if($venta->user_id === auth()->id() || auth()->user()->canManageCurrentTenant())
+                                                                @if((int)$venta->user_id === (int)auth()->id() || auth()->user()->canManageCurrentTenant())
                                                                     <button class="btn btn-sm btn-danger"
                                                                         wire:click="$dispatch('confirm-delete', { id: {{ $venta->id }}, message: '¿Está seguro de eliminar la venta #{{ $venta->numero_folio }}?' })"
                                                                         title="Cancelar">
